Revert "MWPW-151936 Aside Tiger Team Enhancements (#2644)"
This reverts commit dfaf39fed9f10cd14986290cd1f96de75ea495a1.
